requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart
python-dotenv
watchfiles
twilio
requests
streamlit==1.34.0
streamlit-option-menu
plotly==5.22.0
pandas==2.2.2
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.38
langchain-google-genai==1.0.10
google-generativeai==0.7.2